Add files via upload
This Python program is a fun and simple Number Guessing Game where a user plays with their friend and the system to manipulate numbers and try to guess a final number. The game involves thinking of a number, adding numbers, and making some simple calculations to reach a surprising result.